refactor: extract type definitions for BlockDatum, StackedBlockDatum, and StackType into a separate chart types file
